+++ lib/moment_dart.dart
@@ -4,6 +4,6 @@
 library moment_dart;
 
 export './src/moment.dart';
-export 'src/localization.dart';
+export './src/localization.dart';
 export './src/localizations/all.dart';
 export './src/formatters/token.dart';